Restructure the project into smaller modules
Create a new find document, to fetch a list of documents based on entity label
